fix legacy behavior when skipMap is nil
In legacy behaviour, we walk every key but we only refresh the TTL
if the current TTL is below a threshold. This check avoids a
read-modify-write for most of the keys. This behavior was lost
when the optimized walk based on map was added.

We should have legacy and optimized behavior in the code base until
we are ready to switchover the prod cluster.

Testing: enhanced test. All existing tests pass

patchset-6: fix walker cmd
patchset-5: re-add the hash package since it is needed. Need to fix
jenkins env.
patchset-4: remove hash package since there shouldn't be any dependency
patchset-3: try removing the hash.cpp to see if jenkins build passes

Change-Id: I691f2c9d093f4a1c5eb1bed617a4e2add4b66627