Allow InsertInode() to overwrite its destination
Since we cannot synchronously notify the kernel when InsertInode()
creates a new file it is difficult for clients to reliably replace a
file multiple times in close succession because they must unlink it,
insert it, unlink it and then insert it again. In the second unlink it
is possible the kernel has not yet been notified of the insert and so
returns an error and the second insertion will then fail with EEXIST.

Overwriting the file simplifies making this operation atomic and
ordered.