Do not convert strings to byteslices in the hot path
These conversions call runtime.stringtoslicebyte() which in turn calls
runtime.mallocgc() and allocates on the heap.

These allocations do not show up in escape analysis and are found by
disassembly of the quantumfsd binary.

The only remaining allocation is in the default case of the writeArg()
function which is converting a string to byteslice, but that is not in
the hot path.