Remove duplicate fields from HardlinkLeg structure
HardlinkLegs already have a backing record that will be used when these
dentries are published to the durable storage. We might as well use that
record for keeping in-memory information about the hardlink legs.

Even though this might potentially be slower (e.g. we now know
record.Filename() causes a heap allocation), it is exactly the same
thing we do for non-hardlink dentries for which we do not keep duplicate
in-memory fields.