Introduce the HardlinkTableImpl type to encapsulate hardlink management
This type is the sole implementer of the HardlinkTable interface and
will eventually be renamed to HardlinkTable once the leaked abstraction
is controlled.

In this patch, we do not even try to contain the implementation details
and complexity of this new type and the clients and whitebox tests still
reach into it and access its private methods. These issues will be fixed
in the future.