Normalizing file must be dirty
When normalizing a file it may or may not be on the dirty queue. If it
is on the dirty queue then we want to wait until the inode itself
flushes before marking it as publishable. If it isn't on the dirty
queue we want to make it publishable immediately. It's racy to check
if an inode is dirty or not, so simply mark it as dirty to ensure that
it will flush and make itself publishable in a consistent manner.