Add an iteration to the WorkspaceNonce type
This iteration will be incremented everytime a new workspaceroot is
published.

This patch will prevent the newer versions of quantumfs to talk to older
versions of wsdb services as it changes the wsdb wire protocol. Therefore,
it should be deployed with care.

The iteration is not stored in ether yet, therefore, does not require a
wsdb wipe yet.