Assert on recursive RLock calls
The golang sync.RwMutex blocks additional readers once an exclusive
lock is waiting. This implies that if a goroutine were to RLock() the
mutex twice it may deadlock with another goroutine calling Lock().

Detect this situation and assert, noting the two places the single
goroutine calls RLock(). The first is reported manually and the second
is reported in the panic output.