Do not check the hardlinktable when reporting direct children
In gut/review/r/1496, in the process of normalizing the directory
inherits the normalized child, but the entry is not removed from the
hardlink table until the HardlinkDelta of the parent directory is
bubbled up to the root.

In this interval, previously the parent would not report the newly
normalized child as a direct child. This meant, with the correct timing,
the parent might have gotten uninstantiated while the child would live on.