Reduce trace verbosity in fillAttrWithDirectoryRecordTracing
fillAttrWithDirectoryRecordTracing() is called for each file each time
ReadDirPlus is called on a directory. This was generating five lines
of tracing for each filename. This is wasteful of qlog space and
presented high overhead, especially in the case of
objectTypeToFileType(), where the cost of the logging vastly
outweighs the execution time of the function itself. For this high
throughput code-path eliminate the redundancy of traces.

This change reduces the time to create a snapshot of a large directory
by 30%.