Add parentGetChildAttr()
This method is intended to replace parentGetChildRecordCopy() as most
of the users immediately use the record to fill out a fuse.Attr
object. Rather than copying the record only to copy the data again,
simply copy the data directly into the output object.

This copying is necessary to avoid races between retrieving the
record, the underlying hardlink being deleted or the record otherwise
being modified, and using the returned record.