Use pathwalk which completely skips directories
Golang 1.10 modified filepath.Walk() such that it now reads the
directory contents prior to calling the walk function on that
directory. This breaks the Walker tests because this new behaviour
doesn't match that of the Walker itself.

Implement a simple path walker which has the same SkipDir semantics as
the Walker library.