Keep the modified entry
Inside the ChildContainer it may be that the record is publishable
without a separate effective record. In that case modifying the entry
will create a new effective record which is modified according to the
attributes, but these modifications are applied to the new effective
record and not the publishable record we have already grabbed. Thus
the attribute data we pass back to FUSE will be out of date.

Resolve this by re-fetching the record.