Sort by cheap fields first
Accessing the filename from an EncodedDirectoryRecord is somewhat
expensive as it allocates memory for the string. As we only require
that the sort order be consistent we can first sort on fields which
are cheaper to access prior to falling back on the only guaranteed
unique field in a record.