removed TODO
The early return from a DE chain walk is fine when the walk function
returns a SkipEntry on the first DE. All unique keys get walked anyway
if they are not in the cache, once per refresh cycle. So all chained
DEs will get added to cache once. Wether the keys stay in cache within
a refresh cycle (due to cache sizing) doesn't really impact the refresh
requirements.

In summary, its safe to do an early return from chained DE walk when
the first DE is present in cache.