If the FileIds match, the ContentTime will match
If the FileIds of two records within the same directory match, then
they are hardlinked together and the ContentTimes will also match.
Since this sort is in the quantumfs package we are not aware of the
HardlinkLeg type and so cannot compare the hardlink creationTimes.
Thus the only other unique field is the filename.