VeryLargeFile remainders are LargeFiles, not blocks
VeryLargeFiles are composed of a number of LargeFiles. Thus the size
remainder for the last part is a portion of a maximum LargeFile size,
not a portion of a maximum block size. Using the latter results in the
file being truncated to a length which is some multiple of the block
size too small.