Update the ctime of a moving file consistently
The ctime of a moving file must be updated in a consistent matter. There
are some conditions:
1. If the inode is dirty, an effective record can be created with the
   updated ctime.
2. if the inode is instantiated but not dirty, it will be marked as dirty.
3. If the inode is not instantiated, the record will be made publishable
   immediately in either child container or the hardlink table.

Formerly, in case of a rename, we were only updating the name of the
record and were making it publishable immediately. That runs into the
issue of making the changes corresponding to a dirty inode publishable
immediately and consequently publishing an inconsistent image.