Fix the increment/decrement of lookupCount in lookupInternal
Using the value of 0 does not guarantee anything and the inode on which
increaseLookupCount() has been called might be uninstantiated. I am
not sure why 0 was used instead of 1 which is more natural.

Also, do not forget the workspaceroot if the file is an immediate child.
This incorrect bookkeeping was not caught before because deducting
lookupcount by 0 is a no-op.