Fix a race condition the flusher and the forget threads
The flusher has to mark the inode being flushed as clean prior to
calling its flush method so that it will be queued again if it is
dirtied again.

However, a clean inode might be uninstantiated before it gets flushed.
Such a race can only happen for directories that get uninstantiated as
part of the chain starting from one of their heirs.

Incrementing the lookupCount will result in aborting the uninstantiation
of the to-be-flushed inode in the forget thread. The uninstantiation
will still happen in the flusher thread if the inode has to be
uninstantiated.