Track the nonce when skipping flushing
If a workspace is deleted and them immediately recreated we'll be
racing against the workspace DB for the notification of the deletion
to arrive. We must ensure the nonce matches or we'll mark the new
dirty queue as deleted because the deletion notification came after
the workspace was created locally.