Support unique request IDs in Qlogger
Qlogger is a small utility type which works much like daemon.ctx to
make logging easy, but is intended to be used by external users. In
order to make this useful we need to support unique request IDs.

This change also adds an option to set the qlog size instead of using
the default.