qfsclient: Fall back to system tmpdir if ROOTDIRNAME unset
If ROOTDIRNAME is set, use a tempdir under /dev/shm/...
but fall back to using the system tempdir if it isn't set.

use realpath to canonicalize the temp dir name so that
we don't need partial string comparisons to check the API
path - important on macOS which has a symlink in the
system tempdir path.

This change allows the client unit tests to work on macOS.