Add API for bulk lookup count increments
In ReadDirPlus() many entries are returned and each one requires its
lookup count be increased. Logging a trace message and grabbing the
lookupCountLock separately for each of these consumes about 70% of the
total execution time of the function. Amortizing these costs amongst
the many entries reduces the total cost to about 30%.