Zero counts in lookupCounts are invalid
Previously we used zero counts in the lookupCounts table to indicate
that an inode was instantiated, but not referenced by the kernel as a
note that we would need to uninstantiate it at some point. Now that
each Inode is reference counted they are uninstantiated as soon as the
last reference is gone. Therefore zero reference counts are
unnecessary and should not be maintained.