Start with a refcount of 1
For debugging reasons we want to panic when we increment a refcount
from zero, but that means that we need to start with a reference of
some sort. I think we might be able to get away with not grabbing a
reference for every call to qfs.inode(), so we initialize with a
reference we assume we are going to give to the lookupCounts. If we
find that we aren't doing that, remove the reference.