Forget inline
Now that uninstantiation is triggered by reference counts there is no
need to delay processing Forget requests and it is difficult because
we won't know that we may uninstantiate until we have already done so
after returning from shouldForget().

For the moment process Forget inline instead of delaying to a separate
goroutine.