Move refcount into a table atomic with inode map
Normally refcounting will have every reference take a count, this
includes indices such as the inode map. Unfortunately we need to use
reference counting to determine when to remove from the inode map.

Facilitate this by moving the refcounts from an atomic member of the
inode objects themselves into a table parallel to the inode map and
protected by the same lock. Thus when the refcount reaches zero we are
able to atomically remove it from the map. This resolves races like:

    Mux::Forget
    D: 0 refs on inode 10
    Uninstantiating inode 10
                                Mux::Lookup testFile2
                                Mux::incrementLookupCount
                                A: 1 refs on inode 10
    Clearing inode 10