Use local SkipMap for each Walk.
When a walk fails midway the rootID of that workspace still ends up in
the SkipMap. Due to this - in the next iteration - the workspace is
skipped and considered successful.

To solve this, we now maintain a separate map for each walk and merge it to
the global map only if and when the walk succeeds.

Additional changes
-----------------
Signature of SkipMap now uses string for keys, so that SkipMap
API itself can be used in implementing Merge().

Remove walkerCtx from the signature of SkipMap API, since it served no
purpose.

Add files to gitignore

Change-Id: If6c481aa88072ff174f94a16359c86d02a206674