walker: add option for naming walker daemon instance
Currently and most probably in future, we may run multiple
instances of walker, on same/different hosts and each instance
handling different workspaces. This change, adds support for
"naming" a walker daemon instance so that all the stats are
appropriately tagged using "host" (on which the daemon is
run) and "name" (walker daemon instance name).

Change-Id: I54c74b8e364dd563ae8694424887599554424c99
Testing: manual on dev cluster