Fix Structure.md to look alright
This is a bit of a hack. Markdown doesn't seem to provide a nice way
to have definitions or to insert a newline. Instead I append a colon
to the package name. When read as plain text this looks file. When
read as processed markdown it'll be slightly awkward, but still have a
useful divider.